+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Removed
 - Redundant pagination data
+### Fixed
+- Allow creating campaign draft without creatives
 
 ## [2.2.3] - 2022-12-19
 ### Added